Epic 1 Polish: Architectural Debt Remediation
Removes Pydantic V2 discriminator warning from single-item union and decouples domain-specific Canvas/SciVis logic from pure MCP infrastructure adapter.
