+++ src/coreason_manifest/adapters/mcp/__init__.py
@@ -1,22 +1,16 @@
 from coreason_manifest.adapters.mcp.server import (
-    MCPOperation,
-    MCPOperationSequence,
     MCPPromptRef,
     MCPResourceList,
     MCPServerConfig,
-    MCPToolName,
     MCPTransport,
     SSETransportConfig,
     StdioTransportConfig,
 )
 
 __all__ = [
-    "MCPOperation",
-    "MCPOperationSequence",
     "MCPPromptRef",
     "MCPResourceList",
     "MCPServerConfig",
-    "MCPToolName",
     "MCPTransport",
     "SSETransportConfig",
     "StdioTransportConfig",
